abstracted out cleanPhoneNumber function into client/utils/index.ts
